+++ hpproj/hp_helper.py
@@ -9,16 +9,20 @@
 
 from __future__ import print_function, division
 
+from itertools import repeat, product
+
 import logging
 import numpy as np
 import healpy as hp
 
 from astropy.io import fits
-from astropy.wcs import utils as wcs_utils
-from astropy.coordinates import UnitSphericalRepresentation
+from astropy.wcs import WCS, utils as wcs_utils
+from astropy.table import Table, Column
+from astropy.coordinates import SkyCoord, UnitSphericalRepresentation
 from astropy import units as u
 
-from .wcs_helper import equiv_celestial, build_wcs
+from .wcs_helper import equiv_celestial, build_wcs, rot_frame
+from .wcs_helper import build_wcs_profile, build_wcs_cube
 from .wcs_helper import DEFAULT_SHAPE_OUT
 
 from .decorator import _hpmap
@@ -26,7 +30,9 @@ from .decorator import _hpmap
 logging.basicConfig(format='%(asctime)s -- %(levelname)s: %(message)s', level=logging.DEBUG)
 
 __all__ = ['hp_is_nest', 'hp_celestial', 'hp_to_wcs', 'hp_to_wcs_ipx',
-           'hp_project', 'gen_hpmap', 'build_hpmap', 'hpmap_key']
+           'hp_project', 'gen_hpmap', 'build_hpmap', 'hpmap_key',
+           'wcs_to_profile', 'hp_to_profile', 'hp_profile', 'hp_stack',
+           'hp_to_aperture', 'hp_photometry']
 
 
 def hp_celestial(hp_header):
@@ -66,9 +72,11 @@ def hp_is_nest(hp_header):
     Parameters
     ----------
     hp_header : :class:`~astropy.io.fits.header.Header`
-        the header 100
+        the header
+
+    Returns
     -------
-    bool :
+    boolean :
         True if the header is nested
     """
 
@@ -115,6 +123,96 @@ def rotate_frame(alon, alat, hp_header, wcs):
         alat = coords.data.lat.deg
 
     return alon, alat
+
+
+def wcs_to_profile(hdu, wcs, shape_out=DEFAULT_SHAPE_OUT[0]):
+    """Centered profile from 2D map
+
+    Parameters
+    ----------
+    hdu : :class:`astropy.fits.ImageHDU`
+        hdu containing the 2D array and corresponding header, the profile will be made from the CRVAL position
+    wcs : :class:`astropy.wcs.WCS`
+        wcs object to describe the radius of the profile
+    shape_out : int
+        shape of the output profile
+
+    Returns
+    -------
+    :class:`astropy.fits.ImageHDU`
+        1D hdu image containing the profile and the corresponding header
+    """
+
+    frame = wcs_utils.wcs_to_celestial_frame(WCS(hdu.header)).name
+    coord = SkyCoord(hdu.header['CRVAL1'], hdu.header['CRVAL2'], unit="deg", frame=frame)
+
+    shape = hdu.shape
+    yy, xx = np.indices(shape)
+    lon_arr, lat_arr = WCS(hdu.header).wcs_pix2world(xx, yy, 0)
+
+    coords = SkyCoord(lon_arr, lat_arr, unit="deg", frame=frame)
+    dist = coords.separation(coord).to(u.deg).value
+
+    r_edge = wcs.all_pix2world(np.arange(shape_out + 1) - 0.5, 0)[0]
+
+    hist, bin_edges = np.histogram(dist, bins=r_edge, weights=hdu.data)
+    hist_d, bin_edges_d = np.histogram(dist, bins=r_edge)
+
+    with np.errstate(invalid='ignore'):
+        hist /= hist_d
+
+    return hist
+
+
+@_hpmap
+def hp_to_profile(hp_hdu, wcs, coord, shape_out=DEFAULT_SHAPE_OUT[0], std=False):
+    """Extract radial profile from healpix map
+
+    Parameters
+    ----------
+    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+        a pseudo ImageHDU with the healpix map and the associated header
+    wcs : :class:`astropy.wcs.WCS`
+        wcs object to describe the radius of the profile
+    coord : :class:`astropy.coordinate.SkyCoord`
+        the sky coordinate of the center of the profile
+    shape_out : int
+        shape of the output profile
+    std : bool
+        return the standard deviation
+
+    Returns
+    -------
+    :class:`astropy.fits.ImageHDU`
+        1D hdu image containing the profile and the corresponding header,
+        optionnaly a second ImageHDU containing the standard deviation
+    """
+
+    coord = rot_frame(coord, hp_hdu.header['COORDSYS'])
+
+    r_edge = wcs.all_pix2world(np.arange(shape_out + 1) - 0.5, 0)[0]
+
+    nside = hp_hdu.header['NSIDE']
+    nest = hp_is_nest(hp_hdu.header)
+
+    # TODO : can we have a product here...
+    vec = hp.rotator.dir2vec(coord.data.lon.deg, coord.data.lat.deg, lonlat=True)
+
+    # TODO: Another possibility would be to use the outer ring, compute all distances and use np.histogram
+    i_pix = [hp.query_disc(nside, vec, np.radians(radius), nest=nest, inclusive=False) for radius in r_edge]
+
+    # Only keep the pixels why are only present in the outer
+    # radius
+    a_pix = []
+    for in_pix, out_pix in zip(i_pix[:-1], i_pix[1:]):
+        a_pix.append(np.setxor1d(in_pix, out_pix))
+
+    profile = np.asarray([np.mean(hp_hdu.data[pix]) for pix in a_pix])
+    if not std:
+        return profile
+    else:
+        std_profile = np.asarray([np.std(hp_hdu.data[pix]) for pix in a_pix])
+        return profile, std_profile
 
 
 @_hpmap
@@ -229,7 +327,110 @@ def hp_to_wcs_ipx(hp_header, wcs, shape_out=DEFAULT_SHAPE_OUT):
 
 
 @_hpmap
-def hp_project(hp_hdu, coord, pixsize=0.01, npix=512, order=0, projection=('GALACTIC', 'TAN')):
+def hp_project(hp_hdu, coord, pixsize=0.01, shape_out=DEFAULT_SHAPE_OUT, order=0, projection=('GALACTIC', 'TAN')):
+    """Project an healpix map at a single given position
+
+    Parameters
+    ----------
+    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+        a pseudo ImageHDU with the healpix map and the associated header to be projected
+    coord : :class:`astropy.coordinate.SkyCoord`
+        the sky coordinate of the center of the projection
+    pixsize : float
+        size of the pixel (in degree)
+    shape_out : tuple
+        shape of the output map (n_y, n_x)
+    order : int (0|1)
+        order of the interpolation 0: nearest-neighbor, 1: bi-linear interpolation
+    projection : tuple of str
+        the coordinate ('GALACTIC', 'EQUATORIAL') and projection ('TAN', 'SIN', 'GSL', ...) system
+
+    Returns
+    -------
+    :class:`astropy.io.fits.PrimaryHDU`
+        containing the array and the corresponding header
+    """
+
+    proj_sys, proj_type = projection
+
+    wcs = build_wcs(coord, pixsize, shape_out=shape_out, proj_sys=proj_sys, proj_type=proj_type)
+    proj_map = hp_to_wcs(hp_hdu, wcs, shape_out=shape_out, order=order)
+
+    return fits.ImageHDU(proj_map, wcs.to_header(relax=0x20000))
+
+
+@_hpmap
+def hp_stack(hp_hdu, coords, pixsize=0.01, shape_out=DEFAULT_SHAPE_OUT, order=0, projection=('GALACTIC', 'TAN'), keep=False):
+    """Perform stacking on an healpix map
+
+    Parameters
+    ----------
+    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+        a pseudo ImageHDU with the healpix map and the associated header to be projected
+    coords : list of :class:`astropy.coordinate.SkyCoord`
+        list of sky coordinates for the center of the cropped maps
+    pixsize : float
+        size of the pixel (in degree)
+    shape_out : tuple
+        shape of the output map (n_y, n_x)
+    order : int (0|1)
+        order of the interpolation 0: nearest-neighbor, 1: bi-linear interpolation
+    projection : tuple of str
+        the coordinate ('GALACTIC', 'EQUATORIAL') and projection ('TAN', 'SIN', 'GSL', ...) system
+    keep : boolean (default False)
+        return all the cropped maps as a 3D cube instead of one stack map
+
+    Returns
+    -------
+    `:class:~fits.ImageHDU`
+        hdu containing the stack image or cube and corresponding header
+    """
+
+    proj_sys, proj_type = projection
+
+    if coords.isscalar:
+        coords = SkyCoord([coords.frame])
+
+    coords = rot_frame(coords, proj_sys)
+
+    if isinstance(pixsize, (int, float)):
+        pixsize = repeat(pixsize, coords.shape[0])
+    else:
+        pixsize = iter(pixsize)
+
+    w = build_wcs_cube(coords[0], 0, pixsize=next(pixsize), shape_out=shape_out,
+                       proj_sys=proj_sys, proj_type=proj_type)
+
+    # First image outside of the loop
+    _w = w.dropaxis(2)
+    stacks = hp_to_wcs(hp_hdu, _w, shape_out)
+
+    if keep:
+        # return 3D array
+        stacks = np.ma.resize(stacks, coords.shape + stacks.shape)
+
+    for index, (coord, pix) in enumerate(zip(coords[1:], pixsize), 1):
+        _w.wcs.crval = (coord.data.lon.deg, coord.data.lat.deg)
+        _w.wcs.cdelt = [-pix, pix]
+
+        if keep:
+            stacks[index, :, :] = hp_to_wcs(hp_hdu, _w, shape_out)
+        else:
+            stacks += hp_to_wcs(hp_hdu, _w, shape_out)
+
+    # Take the mean
+    if not keep:
+        stacks /= coords.shape[0]
+        w = w.dropaxis(2)
+
+    if coords.shape[0] > 1:
+        w.wcs.crval[0:2] = [0, 0]
+
+    return fits.ImageHDU(stacks, w.to_header(relax=0x20000))
+
+
+@_hpmap
+def hp_profile(hp_hdu, coord, pixsize=0.01, npix=512):
     """Project an healpix map at a single given position
 
     Parameters
@@ -242,12 +443,6 @@ def hp_project(hp_hdu, coord, pixsize=0.01, npix=512, order=0, projection=('GALA
         size of the pixel (in degree)
     npix : int
         number of pixels in the final map, the reference pixel will be at the center
-    order : int (0|1)
-        order of the interpolation 0: nearest-neighbor, 1: bi-linear interpolation
-    projection : tuple of str
-        the coordinate ('GALACTIC', 'EQUATORIAL') and projection ('TAN', 'SIN', 'GSL', ...) system
-    hdu : bool
-        return a :class:`astropy.io.fits.PrimaryHDU` instead of just a ndarray
 
     Returns
     -------
@@ -255,12 +450,91 @@ def hp_project(hp_hdu, coord, pixsize=0.01, npix=512, order=0, projection=('GALA
         containing the array and the corresponding header
     """
 
-    proj_sys, proj_type = projection
+    wcs = build_wcs_profile(pixsize)
+    profile = hp_to_profile(hp_hdu, wcs, coord, shape_out=npix, std=False)
 
-    wcs = build_wcs(coord, pixsize, shape_out=(npix, npix), proj_sys=proj_sys, proj_type=proj_type)
-    proj_map = hp_to_wcs(hp_hdu, wcs, shape_out=(npix, npix), order=order)
+    return fits.ImageHDU(profile, wcs.to_header(relax=0x20000))
 
-    return fits.PrimaryHDU(proj_map, wcs.to_header(relax=0x20000))
+
+@_hpmap
+def hp_to_aperture(hp_hdu, coords, apertures=None):
+    """Raw aperture summation on an healpix map
+
+    Parameters
+    ----------
+    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+        a pseudo ImageHDU with the healpix map and the associated header to be projected
+    coords : list of :class:`astropy.coordinate.SkyCoord`
+        the sky coordinates for the center of the apertures
+    apertures : list of `:class:astropy.coordinates.Angles`
+        aperture angle in which we perfom summation
+
+    Returns
+    -------
+    npix, apertures : array_like
+        2 arrays containing the number of pixels, and sum of the pixels
+        within the aperture respectively
+    """
+
+    coords = rot_frame(coords, hp_hdu.header['COORDSYS'])
+
+    nside = hp_hdu.header['NSIDE']
+    nest = hp_is_nest(hp_hdu.header)
+
+    vecs = hp.rotator.dir2vec(coords.data.lon.deg, coords.data.lat.deg, lonlat=True)
+
+    if coords.isscalar or coords.shape == (1,):
+        vecs = vecs[:, np.newaxis]
+
+    pix_indexes = [hp.query_disc(nside, vec, np.radians(radius.degree), nest=nest, inclusive=False)
+                   for vec, radius in product(vecs.T, apertures)]
+
+    # TODO: have a look at regions.PixelRegion for exact boundings
+    # apertures = [(len(index), np.sum(hp_hdu.data[index])) if len(index) > 0 else None for index in pix_indexes]
+    npix, apertures = zip(*[(index.shape[0], np.sum(hp_hdu.data[index])) for index in pix_indexes])
+
+    return np.reshape(npix, coords.shape + (-1,)), np.reshape(apertures, coords.shape + (-1,))
+
+
+@_hpmap
+def hp_photometry(hp_hdu, coords, apertures):
+    """Aperture photometry on an healpix map at a single given position
+
+    Parameters
+    ----------
+    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+        a pseudo ImageHDU with the healpix map and the associated header to be projected
+    coords : :class:`astropy.coordinate.SkyCoord`
+        the sky coordinates for the center of the apertures
+    apertures : 3 `:class:astropy.coordinates.Angles`
+        3 floats defining the aperture radius and inner/outer annulus radii
+
+    Returns
+    -------
+    :class:`astropy.io.fits.BinaryHDU`
+        table containing the photometry
+    """
+
+    if isinstance(coords, list):
+        coords = SkyCoord(coords)
+    elif coords.isscalar:
+        coords = SkyCoord([coords])
+
+    npix, S = hp_to_aperture(hp_hdu, coords, apertures)
+
+    S_aper, S_inner, S_outer = S.T
+    n_aper, n_inner, n_outer = npix.T
+
+    # Perfom photometry
+    mean_background = (S_outer - S_inner) / (n_outer - n_inner)
+    mean_brigthness = S_aper / n_aper - mean_background
+
+    result = Table()
+    result['brigthness'] = Column(mean_brigthness, unit=hp_hdu.header['UNIT'])
+    result['background'] = Column(mean_background, unit=hp_hdu.header['UNIT'])
+    result['n_pix'] = Column(n_aper)
+
+    return result
 
 
 def gen_hpmap(maps):