+++ hpproj/visu.py
@@ -32,7 +32,7 @@ def view(hp_hdu, coord=None, npix=360, proj_sys='GALACTIC', proj_type='TAN', asp
 
     Parameters
     ----------
-    hp_hdu : `:class:astropy.io.fits.ImageHDU`
+    hp_hdu : :class:`~astropy.io.fits.ImageHDU`
         a pseudo ImageHDU with the healpix map and the associated header to be projected
     coord : :class:`astropy.coordinate.SkyCoord`
         the sky coordinate of the center of the projection
@@ -65,7 +65,7 @@ def view(hp_hdu, coord=None, npix=360, proj_sys='GALACTIC', proj_type='TAN', asp
     return ImageHDU(_data, _wcs.to_header())
 
 
-mollview = partial(view, proj_type='TAN', aspect=0.5)
+mollview = partial(view, proj_type='MOL', aspect=0.5)
 update_wrapper(mollview, view)
 mollview.__name__ = "mollview"
 mollview.__doc__ = "Mollweide " + mollview.__doc__