+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 master
 ======
 
+Added
+-----
+* Profile routine
+* hp_project use shape_out instead of npix
+* Add hp_stack to perform stacking on healpix map
+* Add hp_profile to extract profile at given SkyCoord
+* Add hp_photometry to perform aperture photometry on list of SkyCoord
+
 Fixed
 -----
 * Refactoring : split wcs_helper from hp_helper